a card title can now be specified in the prompt
maincards.py:
Used claude to update the code to enable the model to pass a title. Command line Prompt example:
python maincards.py "generate a pdf card deck with a comma separated list of 50 unique verbs with a title of Verbs"

genenate_card_pdf.py: updated to use the new title parameter to be the title of each card and the name of the pdf.  Also remove a remaining "noun" reference.