Now using config files and chat interface without tool calls
I learned that the model doesn't expect to give its entire list of strings to the tool when calling it.  It crops the list to 3-10, even though after the tool call, it generates the entire list of 100-150 items.   So I switched to using a config file for the prompt and the PDF settings.  Now I call the model first and get the entire list of items and pass that to the PDF generator.
I also needed to split the system prompt into two for llama3.1, so that I tailor the prompt more clearly JUST for questions and JUST for word lists.  That made the model's list output more consistent, though not perfectly consistent.  I also had to put in a lot of string trimming and cleaning code to ensure the list of questions was properly formatted.  It does well 9 times out of 10 right now.

TODO: Later I need to move that code outside of the pdf generator and into the code that calls the model.  That way, when I try different models, I can tailor the cleaning code to the model's output, as I expect each model will behave differently.

The correct python file to call is currently called:
mainchatgencards.py

I'll be moving that to main. after this commit.