UI Content Type now supports dropdowns with 3 options
UI Content Type now supports dropdowns with 3 options:
"Words", "Questions", "QuestionsAndAnswers"

I added a new config type for cards that will eventually show both the question and the answer upside down on the same card.  For now, the QuesstionsAndAnswers datatype will cause the rest of the logic to fail.  So with this commit, only use Words and Questions. in the config files.