now supports Questions and Answers on the same card
generate_card_pdf.py: updated to handle questions and answers as alternating lines from llama. puts each answer at bottom of same card as question.  move the "is question?" logit to main, to make this code more generic

main.py: updated prompts to llama3.1 to request questions and answers option.   move the string cleaning and splitting into an array from llama's out put into this file.  Previously the list of output items was passed as a string, now it is an array.